+++ src/Data/Singletons/Internal.hs
@@ -3,7 +3,7 @@
              TypeFamilies, TypeOperators, TypeFamilyDependencies,
              UndecidableInstances, TypeInType, ConstraintKinds,
              ScopedTypeVariables, TypeApplications, AllowAmbiguousTypes,
-             PatternSynonyms, ViewPatterns #-}
+             PatternSynonyms, ViewPatterns, MultiParamTypeClasses #-}
 
 -----------------------------------------------------------------------------
 -- |
@@ -50,17 +50,15 @@ data family Sing (a :: k)
 
 -- | A 'SingI' constraint is essentially an implicitly-passed singleton.
 -- If you need to satisfy this constraint with an explicit singleton, please
--- see 'withSingI' or pattern match using the 'SingI' pattern synonym.
+-- see 'withSingI' or the 'SingI' pattern synonym.
 class SingI (a :: k) where
   -- | Produce the singleton explicitly. You will likely need the @ScopedTypeVariables@
   -- extension to use this method the way you want.
   sing :: Sing a
 
--- | Pattern matching on an explicit @Sing a@ to get an implicit
--- @SingI a@ constraint.
---
--- /Since: 2.4/
-pattern SingI :: forall (a::k). () => SingI a => Sing a
+-- | Permits pattern matching on an explicit @Sing a@ value to bring an
+-- implicit @SingI a@ constraint into scope.
+pattern SingI :: forall (a :: k). () => SingI a => Sing a
 pattern SingI <- (singInstance -> SingInstance)
   where SingI = sing
 
@@ -85,6 +83,24 @@ class SingKind k where
 
   -- | Convert an unrefined type to an existentially-quantified singleton type.
   toSing   :: Demote k -> SomeSing k
+
+-- | @SingKindOf demoted k@ is shorthand for @(SingKind k, Demote k ~
+-- demoted)@.
+--
+-- It lets us write 'projSigma1' from @Data.Singletons.Sigma@ can be written
+--
+-- @
+-- demote :: forall a. (SingKindOf demoted (KindOf a), SingI a) => demoted
+-- demote = fromSing (sing @(KindOf a) @a)
+-- @
+--
+-- instead of
+--
+-- @
+-- demote :: forall a. (SingKind (KindOf a), SingI a) => Demote (Demote (KindOf a))
+-- @
+class    (SingKind k, Demote k ~ demoted) => SingKindOf demoted k
+instance (SingKind k, Demote k ~ demoted) => SingKindOf demoted k
 
 -- | An /existentially-quantified/ singleton. This type is useful when you want a
 -- singleton type, but there is no way of knowing, at compile-time, what the type
@@ -288,9 +304,9 @@ withSingI sn r =
 
 -- | Convert a normal datatype (like 'Bool') to a singleton for that datatype,
 -- passing it into a continuation.
-withSomeSing :: forall k r
-              . SingKind k
-             => Demote k                          -- ^ The original datatype
+withSomeSing :: forall k r demoted
+              . SingKindOf demoted k
+             => demoted                           -- ^ The original datatype
              -> (forall (a :: k). Sing a -> r)    -- ^ Function expecting a singleton
              -> r
 withSomeSing x f =
@@ -308,8 +324,8 @@ withSing f = f sing
 -- property.  If the singleton does not satisfy the property, then the function
 -- returns 'Nothing'. The property is expressed in terms of the underlying
 -- representation of the singleton.
-singThat :: forall (a :: k). (SingKind k, SingI a)
-         => (Demote k -> Bool) -> Maybe (Sing a)
+singThat :: forall (a :: k) demoted. (SingKindOf demoted k, SingI a)
+         => (demoted -> Bool) -> Maybe (Sing a)
 singThat p = withSing $ \x -> if p (fromSing x) then Just x else Nothing
 
 -- | Allows creation of a singleton when a proxy is at hand.
@@ -331,5 +347,5 @@ singByProxy# _ = sing
 --
 -- >>> demote @(Nothing :: Maybe Ordering)
 -- Nothing
-demote :: forall a. (SingKind (KindOf a), SingI a) => Demote (KindOf a)
+demote :: forall a demoted. (SingKindOf demoted (KindOf a), SingI a) => demoted
 demote = fromSing (sing @(KindOf a) @a)